+++ src/Reporter.zig
@@ -106,6 +106,7 @@ pub const Error = enum(u8) {
     inferred_type = 90,
     empty_import = 91,
     import_already_exists = 92,
+    code_after_return = 93,
 };
 
 // Inspired by https://github.com/zesterer/ariadne